move partial with docstring
