+++ qcodes/instrument_drivers/rigol/DG1062.py
@@ -4,7 +4,7 @@ from typing import Any, Dict, List, Union, cast
 
 from qcodes import ChannelList, InstrumentChannel, VisaInstrument
 from qcodes import validators as vals
-from qcodes.utils.helpers import partial_with_docstring
+from qcodes.utils import partial_with_docstring
 
 log = logging.getLogger(__name__)
