+++ qcodes/utils/helpers.py
@@ -50,6 +50,7 @@ from .deep_update_utils import deep_update
 from .full_class import full_class
 from .function_helpers import is_function
 from .json_utils import NumpyJSONEncoder
+from .partial_utils import partial_with_docstring
 from .path_helpers import QCODES_USER_PATH_ENV, get_qcodes_path, get_qcodes_user_path
 from .qt_helpers import foreground_qt_window
 from .spyder_utils import add_to_spyder_UMR_excludelist
@@ -84,34 +85,3 @@ def attribute_set_to(object_: object,
         yield
     finally:
         setattr(object_, attribute_name, old_value)
-
-
-def partial_with_docstring(func: Callable[..., Any],
-                           docstring: str,
-                           **kwargs: Any) -> Callable[..., Any]:
-    """
-    We want to have a partial function which will allow us to access the docstring
-    through the python built-in help function. This is particularly important
-    for client-facing driver methods, whose arguments might not be obvious.
-
-    Consider the follow example why this is needed:
-
-    >>> from functools import partial
-    >>> def f():
-    >>> ... pass
-    >>> g = partial(f)
-    >>> g.__doc__ = "bla"
-    >>> help(g) # this will print the docstring of partial and not the docstring set above
-
-    Args:
-        func: A function that its docstring will be accessed.
-        docstring: The docstring of the corresponding function.
-    """
-    ex = partial(func, **kwargs)
-
-    def inner(*inner_args: Any, **inner_kwargs: Any) -> Any:
-        return ex(*inner_args, **inner_kwargs)
-
-    inner.__doc__ = docstring
-
-    return inner
